+++ src/client/about-dialog/Dialog.svelte
@@ -1,12 +1,13 @@
 <script lang="ts">
+  import { A } from 'flowbite-svelte';
   import Application from '../Application.svelte';
 </script>
 
 <Application>
-  <p class="mb-4 text-green-400">Thank you for using Firesheet</p>
-  <p>The code for this Google Apps Script can be found here:</p>
-  <a target="_blank" href="https://github.com/melledijkstra/gas-fire">
+  <p class="mb-4">Thank you for using Firesheet</p>
+  <p class="mb-4">The code for this Google Apps Script can be found here:</p>
+  <A class="mb-4" target="_blank" href="https://github.com/melledijkstra/gas-fire">
     @melledijkstra/gas-fire
-  </a>
+  </A>
   <p>{`Version: ${__APP_VERSION__}`}</p>
 </Application>