+++ src/client/components/Test.svelte
@@ -1,5 +1,0 @@
-<script lang="ts">
-
-</script>
-
-<p class="text-green-600 text-5xl">Test Component!!</p>