+++ src/client/import-dialog/index.ts
@@ -1,6 +1,0 @@
-import { mount } from 'svelte';
-import ImportDialog from './Dialog.svelte';
-
-mount(ImportDialog, {
-  target: document.getElementById('index')!
-})