+++ src/client/settings-dialog/index.ts
@@ -1,6 +1,0 @@
-import { mount } from 'svelte';
-import { Dialog } from './Dialog.svelte';
-
-mount(Dialog, {
-  target: document.getElementById('index')!
-})